moved preview component & fixed select
