+++ src/modules/HomePage/components/Preview/Preview.module.scss
